+++ CHANGELOG.md
@@ -12,6 +12,7 @@ https://github.com/jcorporation/myMPD/
 - Feat: Fail gracefully if API request fails or is redirected #984
 - Feat: improve queue sort handling #983
 - Feat: use offcanvas instead of popover for contextmenus #974
+- Feat: better compression of webfont #991
 - Upd: bootstrap.native 5.0.4
 
 ***